server/requirements.txt
boto3==1.34.95
fastapi==0.110.3
python-dotenv==1.0.1
uvicorn==0.29.0
sqlalchemy==2.0.0
aiosqlite==0.19.0